modify
git-svn-id: https://svn.baidu.com/myspace/sep/chenghuige/melt-train@124500 c4ccc3fd-b13d-49bf-a705-f618d8f4424a

+++ include/Trainers/FastRank/FeatureHistogram.h
@@ -66,7 +66,7 @@ namespace gezi {
 			{
 				THROW("length of input arrays is inconsistent");
 			}
-			if (numBins > sumTarget.size())
+			if (numBins > (int)sumTarget.size())
 			{
 				THROW("number of bins is greater than length of input array");
 			}
